Comms handling refactor : - Massive despaghettification, retsructured into a single class hierarchy - Objects made non volatile, should be much more memory efficient. - Transmitters reevaluation is now only triggered from the VesselData modification event, not from callbacks anymore. - A single CommHandler per vessel, created, referenced and updated from VesselData - CommHandler is the base class, does almost nothing in itself (only the bare minimum needed for API handlers) - CommHandler base is responsible for calling the API handler, if one exists - Derivatives from CommHandler for CommNet (vessels / serenity experiments) and RT - All CommHandler types manipulate a single, non volatile, immutable ConnectionInfo, data only class accessed from VesselData - Commented the "wait for 2 seconds" initialization delay, explicit per implementation checks instead. Might need further testing, can be restored if needed. - TODO : The RT implementatin has been transposed without any testing, I will try to give a look. - TODO : implement an editor specific handler

+++ src/Kerbalism/UI/Monitor.cs
@@ -651,20 +651,20 @@ namespace KERBALISM
 			Texture2D image = Textures.signal_red;
 			switch (conn.status)
 			{
-				case LinkStatus.direct_link:
+				case (int)LinkStatus.direct_link:
 					image = conn.strength > 0.05 ? Textures.signal_white : Textures.iconSwitch(Textures.signal_yellow, image);   // or 5% signal strength
 					break;
 
-				case LinkStatus.indirect_link:
+				case (int)LinkStatus.indirect_link:
 					image = conn.strength > 0.05 ? Textures.signal_white : Textures.iconSwitch(Textures.signal_yellow, image);   // or 5% signal strength
 					tooltip += Lib.Color("\n" + Local.UI_Signalrelayed, Lib.Kolor.Yellow);
 					break;
 
-				case LinkStatus.plasma:
+				case (int)LinkStatus.plasma:
 					tooltip += Lib.Color(Lib.Italic("\n" + Local.UI_Plasmablackout), Lib.Kolor.Red);
 					break;
 
-				case LinkStatus.storm:
+				case (int)LinkStatus.storm:
 					tooltip += Lib.Color(Lib.Italic("\n" + Local.UI_Stormblackout), Lib.Kolor.Red);
 					break;
 			}
